backend/requirements.txt
# Web framework
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22

# Imagem
pillow==11.2.1

# HTTP
requests==2.32.5

# Embeddings / busca híbrida
transformers==4.57.6
sentence-transformers==3.4.1
rank-bm25==0.2.2
numpy==1.26.4
deep-translator==1.11.4

azure-storage-blob==12.19.1
azure-identity==1.17.1